Add Android signing keystore property to project file
Added the `<AndroidSigningKeyStore>` property to `SkribeMaui.csproj`
to specify the path to the Android signing key store file
(`signing\my-release-key.keystore`). This change enables the
project to use a specific keystore for signing Android APKs,
a necessary step for preparing the app for release and
distribution on platforms like the Google Play Store.